requirements.txt
astroid==2.3.3
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
Django==2.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.11.2
